requirements.txt
smbus2
pimoroni-bme280
influxdb-client
HAP-python[QRCode]
python-mpd2
falcon